# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Marketing OS Lite API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    put:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /campaigns/{id}/approve:
    post:
      operationId: approveCampaign
      tags: [campaigns]
      summary: Approve a campaign (mark as ready)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/summary:
    get:
      operationId: getCampaignSummary
      tags: [campaigns]
      summary: Dashboard campaign summary
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Campaign summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignSummary"

  # Generated Assets
...
  /connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List platform connections
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlatformConnection"
    post:
      operationId: createConnection
      tags: [connections]
      summary: Connect to a mock ad platform
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConnectionBody"
      responses:
        "201":
          description: Connected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlatformConnection"
  /connections/{id}:
    delete:
      operationId: deleteConnection
      tags: [connections]
      summary: Disconnect a platform
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Disconnected
  /connections/{id}/sync:
    post:
      operationId: syncConnection
      tags: [connections]
      summary: Trigger a mock data sync
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Sync triggered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncJob"

  # Tracking Links
  /tracking-links:
    get:
      operationId: listTrackingLinks
      tags: [tracking]
      summary: List tracking links
      parameters:
        - name: campaignId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of tracking links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TrackingLink"
    post:
      operationId: createTrackingLink
      tags: [tracking]
      summary: Generate a UTM tracking link
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackingLinkBody"
      responses:
        "201":
          description: Tracking link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingLink"
  /tracking-links/{id}:
    delete:
      operationId: deleteTrackingLink
      tags: [tracking]
      summary: Delete a tracking link
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Metrics
...
  /recommendations:
    get:
      operationId: listRecommendations
      tags: [recommendations]
      summary: List recommendations
      parameters:
        - name: workspaceId
          in: query
          schema:
            type: integer
        - name: campaignId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"
  /recommendations/generate:
    post:
      operationId: generateRecommendations
      tags: [recommendations]
      summary: Run rules engine to generate recommendations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                workspaceId:
                  type: integer
              required: [workspaceId]
      responses:
        "200":
          description: Generated recommendations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Recommendation"

  # Audit Log
...
components:
  schemas:
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        name:
          type: string
        objective:
          type: string
          enum: [awareness, traffic, leads, sales]
        productService:
          type: string
        audience:
          type: string
        geography:
          type: string
        budgetSuggestion:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        channels:
          type: array
          items:
            type: string
        landingUrl:
          type: string
        status:
          type: string
          enum: [draft, active, paused, completed, approved]
        createdAt:
          type: string
      required: [id, workspaceId, name, objective, productService, audience, geography, budgetSuggestion, startDate, endDate, channels, landingUrl, status, createdAt]
    CreateCampaignBody:
      type: object
      properties:
        workspaceId:
          type: integer
        name:
          type: string
        objective:
          type: string
          enum: [awareness, traffic, leads, sales]
        productService:
          type: string
        audience:
          type: string
        geography:
          type: string
        budgetSuggestion:
          type: number
        startDate:
          type: string
        endDate:
          type: string
        channels:
          type: array
          items:
            type: string
        landingUrl:
          type: string
      required: [workspaceId, name, objective, productService, audience, geography, budgetSuggestion, startDate, endDate, channels, landingUrl]
    CampaignSummary:
      type: object
      properties:
        total:
          type: integer
        draft:
          type: integer
        active:
          type: integer
        approved:
          type: integer
        completed:
          type: integer
      required: [total, draft, active, approved, completed]
...
    PlatformConnection:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        platform:
          type: string
          enum: [instagram, snapchat, youtube, x]
        accountName:
          type: string
        accountId:
          type: string
        status:
          type: string
          enum: [connected, disconnected, error]
        mockSpend:
          type: number
        mockImpressions:
          type: integer
        mockClicks:
          type: integer
        lastSyncAt:
          type: string
        createdAt:
          type: string
      required: [id, workspaceId, platform, accountName, accountId, status, mockSpend, mockImpressions, mockClicks, createdAt]
    CreateConnectionBody:
      type: object
      properties:
        workspaceId:
          type: integer
        platform:
          type: string
          enum: [instagram, snapchat, youtube, x]
        accountName:
          type: string
      required: [workspaceId, platform, accountName]
    SyncJob:
      type: object
      properties:
        id:
          type: integer
        connectionId:
          type: integer
        status:
          type: string
        startedAt:
          type: string
        completedAt:
          type: string
      required: [id, connectionId, status, startedAt]
    TrackingLink:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        channel:
          type: string
        source:
          type: string
        medium:
          type: string
        campaign:
          type: string
        content:
          type: string
        finalUrl:
          type: string
        generatedTrackingUrl:
          type: string
        createdAt:
          type: string
      required: [id, campaignId, channel, source, medium, campaign, content, finalUrl, generatedTrackingUrl, createdAt]
    CreateTrackingLinkBody:
      type: object
      properties:
        campaignId:
          type: integer
        channel:
          type: string
        source:
          type: string
        medium:
          type: string
        campaign:
          type: string
        content:
          type: string
        finalUrl:
          type: string
      required: [campaignId, channel, source, medium, campaign, content, finalUrl]
...
    Recommendation:
      type: object
      properties:
        id:
          type: integer
        workspaceId:
          type: integer
        campaignId:
          type: integer
        type:
          type: string
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [high, medium, low]
        isRead:
          type: boolean
        createdAt:
          type: string
      required: [id, workspaceId, type, title, description, priority, isRead, createdAt]
...